ldap_exporter/requirements.txt
prometheus_client
python-ldap
